Refactor stripe payment components
